fix(RHINENG-19487): Fix Breadcrumbs in IOP environment
