+++ src/PresentationalComponents/Breadcrumbs/Breadcrumbs.js
@@ -7,6 +7,7 @@ import { useGetRecQuery } from '../../Services/Recs';
 import { Link, useLocation } from 'react-router-dom';
 import { EnvironmentContext } from '../../App';
 import InsightsLink from '@redhat-cloud-services/frontend-components/InsightsLink';
+import { buildBreadcrumbs } from './helpers';
 
 const Breadcrumbs = ({ current }) => {
   const location = useLocation().pathname?.split('/');
@@ -21,40 +22,7 @@ const Breadcrumbs = ({ current }) => {
   );
 
   useEffect(() => {
-    const buildBreadcrumbs = () => {
-      let crumbs = [];
-
-      // add base
-      if (location[3]) {
-        const baseNameWithCapitalLetter =
-          location[3].slice(0, 1).toUpperCase() + location[3].slice(1);
-        crumbs.push({
-          title: `${baseNameWithCapitalLetter}`,
-          navigate: `/${location[3]}`,
-        });
-      }
-
-      // if applicable, add :id breadcrumb
-      if (!skip) {
-        crumbs.push({
-          title: data?.description,
-          navigate: `/${location[1]}/${location[2]}`,
-        });
-      }
-
-      if (location[4] === 'pathways') {
-        crumbs = [
-          {
-            title: 'Pathways',
-            navigate: '/recommendations/pathways',
-          },
-        ];
-      }
-
-      setItems(crumbs);
-    };
-
-    buildBreadcrumbs();
+    setItems(buildBreadcrumbs(location, skip));
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [data]);
